fixing neomake

+++ autoload/neomake/signs.vim
@@ -1,31 +1,33 @@
 " vim: ts=4 sw=4 et
 
-function! s:InitSigns() abort
-    let s:sign_queue = {
-        \ 'project': {},
-        \ 'file': {}
-        \ }
-    let s:last_placed_signs = {
-        \ 'project': {},
-        \ 'file': {}
-        \ }
-    let s:placed_signs = {
-        \ 'project': {},
-        \ 'file': {}
-        \ }
-    let s:sign_queue = {
-        \ 'project': {},
-        \ 'file': {}
-        \ }
-    let s:neomake_sign_id = {
-        \ 'project': {},
-        \ 'file': {}
-        \ }
-endfunction
-call s:InitSigns()
+scriptencoding utf-8
+
+if !has('signs')
+    call neomake#utils#ErrorMessage('Trying to load signs.vim, without +signs.')
+    finish
+endif
+
+let s:sign_queue = {
+    \ 'project': {},
+    \ 'file': {}
+    \ }
+let s:last_placed_signs = {
+    \ 'project': {},
+    \ 'file': {}
+    \ }
+let s:placed_signs = {
+    \ 'project': {},
+    \ 'file': {}
+    \ }
+let s:neomake_sign_id = {
+    \ 'project': {},
+    \ 'file': {}
+    \ }
+
+exe 'sign define neomake_invisible'
 
 " Reset signs placed by a :Neomake! call
-" (resettting signs means the current signs will be deleted on the next call to ResetProject)
+" (resetting signs means the current signs will be deleted on the next call to ResetProject)
 function! neomake#signs#ResetProject() abort
     let s:sign_queue.project = {}
     for buf in keys(s:placed_signs.project)
@@ -97,7 +99,7 @@ function! neomake#signs#PlaceSign(entry, type) abort
 endfunction
 
 function! neomake#signs#CleanAllOldSigns(type) abort
-    call neomake#utils#DebugObject("Removing signs", s:last_placed_signs)
+    call neomake#utils#DebugObject('Removing signs', s:last_placed_signs)
     for buf in keys(s:last_placed_signs[a:type])
         call neomake#signs#CleanOldSigns(buf, a:type)
     endfor
@@ -109,11 +111,13 @@ function! neomake#signs#CleanOldSigns(bufnr, type) abort
         return
     endif
     call neomake#utils#DebugObject('Cleaning old signs in buffer '.a:bufnr.': ', s:last_placed_signs[a:type])
-    for ln in keys(s:last_placed_signs[a:type][a:bufnr])
-        let cmd = 'sign unplace '.s:last_placed_signs[a:type][a:bufnr][ln].' buffer='.a:bufnr
-        call neomake#utils#DebugMessage('Unplacing sign: '.cmd)
-        exe cmd
-    endfor
+    if bufexists(str2nr(a:bufnr))
+        for ln in keys(s:last_placed_signs[a:type][a:bufnr])
+            let cmd = 'sign unplace '.s:last_placed_signs[a:type][a:bufnr][ln].' buffer='.a:bufnr
+            call neomake#utils#DebugMessage('Unplacing sign: '.cmd)
+            exe cmd
+        endfor
+    endif
     unlet s:last_placed_signs[a:type][a:bufnr]
 endfunction
 
@@ -137,26 +141,26 @@ function! neomake#signs#PlaceVisibleSigns() abort
     endfor
 endfunction
 
-exe 'sign define neomake_invisible'
-
-function! neomake#signs#RedefineSign(name, opts)
+function! neomake#signs#RedefineSign(name, opts) abort
     let sign_define = 'sign define '.a:name
     for attr in keys(a:opts)
         let sign_define .= ' '.attr.'='.a:opts[attr]
     endfor
     exe sign_define
 
-    for buf in keys(s:placed_signs)
-        for ln in keys(s:placed_signs[buf])
-            let sign_id = s:placed_signs[buf][ln]
-            exe 'sign place '.sign_id.' name=neomake_invisible buffer='.buf
-            exe 'sign place '.sign_id.' name='.a:name.' buffer='.buf
+    for type in keys(s:placed_signs)
+        for buf in keys(s:placed_signs[type])
+            for ln in keys(s:placed_signs[type][buf])
+                let sign_id = s:placed_signs[type][buf][ln]
+                exe 'sign place '.sign_id.' name=neomake_invisible buffer='.buf
+                exe 'sign place '.sign_id.' name='.a:name.' buffer='.buf
+            endfor
         endfor
     endfor
 endfunction
 
-function! neomake#signs#RedefineErrorSign(...)
-    let default_opts = {'text': '✖'}
+function! neomake#signs#RedefineErrorSign(...) abort
+    let default_opts = {'text': '✖', 'texthl': 'NeomakeErrorSign'}
     let opts = {}
     if a:0
         call extend(opts, a:1)
@@ -167,8 +171,8 @@ function! neomake#signs#RedefineErrorSign(...)
     call neomake#signs#RedefineSign('neomake_err', opts)
 endfunction
 
-function! neomake#signs#RedefineWarningSign(...)
-    let default_opts = {'text': '⚠'}
+function! neomake#signs#RedefineWarningSign(...) abort
+    let default_opts = {'text': '⚠', 'texthl': 'NeomakeWarningSign'}
     let opts = {}
     if a:0
         call extend(opts, a:1)
@@ -179,8 +183,8 @@ function! neomake#signs#RedefineWarningSign(...)
     call neomake#signs#RedefineSign('neomake_warn', opts)
 endfunction
 
-function! neomake#signs#RedefineMessageSign(...)
-    let default_opts = {'text': '➤'}
+function! neomake#signs#RedefineMessageSign(...) abort
+    let default_opts = {'text': '➤', 'texthl': 'NeomakeMessageSign'}
     let opts = {}
     if a:0
         call extend(opts, a:1)
@@ -191,26 +195,60 @@ function! neomake#signs#RedefineMessageSign(...)
     call neomake#signs#RedefineSign('neomake_msg', opts)
 endfunction
 
-function! neomake#signs#RedefineInformationalSign(...)
-    let default_opts = {'text': 'ℹ'}
+function! neomake#signs#RedefineInfoSign(...) abort
+    let default_opts = {'text': 'ℹ', 'texthl': 'NeomakeInfoSign'}
     let opts = {}
     if a:0
         call extend(opts, a:1)
-    elseif exists('g:neomake_informational_sign')
-        call extend(opts, g:neomake_informational_sign)
+    elseif exists('g:neomake_info_sign')
+        call extend(opts, g:neomake_info_sign)
     endif
     call extend(opts, default_opts, 'keep')
     call neomake#signs#RedefineSign('neomake_info', opts)
 endfunction
 
-
-let s:signs_defined = 0
-function! neomake#signs#DefineSigns()
-    if !s:signs_defined
-        let s:signs_defined = 1
-        call neomake#signs#RedefineErrorSign()
-        call neomake#signs#RedefineWarningSign()
-        call neomake#signs#RedefineInformationalSign()
-        call neomake#signs#RedefineMessageSign()
+function! neomake#signs#HlexistsAndIsNotCleared(group) abort
+    if !hlexists(a:group)
+        return 0
     endif
+    return neomake#utils#redir('hi '.a:group) !~# 'cleared'
 endfunction
+
+function! neomake#signs#DefineHighlights() abort
+    let ctermbg = neomake#utils#GetHighlight('SignColumn', 'bg')
+    let guibg = neomake#utils#GetHighlight('SignColumn', 'bg#')
+    let bg = 'ctermbg='.ctermbg.' guibg='.guibg
+
+    for [group, fg_from] in items({
+                \ 'NeomakeErrorSign': ['Error', 'bg'],
+                \ 'NeomakeWarningSign': ['Todo', 'fg'],
+                \ 'NeomakeInfoSign': ['Question', 'fg'],
+                \ 'NeomakeMessageSign': ['ModeMsg', 'fg']
+                \ })
+        let [fg_group, fg_attr] = fg_from
+        let ctermfg = neomake#utils#GetHighlight(fg_group, fg_attr)
+        let guifg = neomake#utils#GetHighlight(fg_group, fg_attr.'#')
+        " Ensure that we're not using SignColumn bg as fg (as with gotham
+        " colorscheme, issue https://github.com/neomake/neomake/pull/659).
+        if ctermfg == ctermbg && guifg == guibg
+            let fg_attr = neomake#utils#ReverseSynIDattr(fg_attr)
+            let ctermfg = neomake#utils#GetHighlight(fg_group, fg_attr)
+            let guifg = neomake#utils#GetHighlight(fg_group, fg_attr.'#')
+        endif
+        exe 'hi '.group.'Default ctermfg='.ctermfg.' guifg='.guifg.' '.bg
+        if !neomake#signs#HlexistsAndIsNotCleared(group)
+            exe 'hi link '.group.' '.group.'Default'
+        endif
+    endfor
+endfunction
+
+function! neomake#signs#DefineSigns() abort
+    call neomake#signs#RedefineErrorSign()
+    call neomake#signs#RedefineWarningSign()
+    call neomake#signs#RedefineInfoSign()
+    call neomake#signs#RedefineMessageSign()
+endfunction
+
+" Init.
+call neomake#signs#DefineHighlights()
+call neomake#signs#DefineSigns()
